step 4 update
